rename: api.js를 apis폴더로 이동
